fix nshots schema

+++ lib/BloqadeSchema/src/types.jl
@@ -61,7 +61,7 @@ end
     rabi_frequency_amplitude_max_slope::Float64 = 0.1
     rabi_frequency_phase_max_slope::Float64 = 0.1
     rabi_detuning_max_slope::Float64 = 0.1
-    n_shots::Float64
+    n_shots::Int
 end
 
 @option struct ShotOutput <: QuEraSchema
